Fix 64-layer support by upgrading layer state bitmasks to uint64_t
The firmware was failing with 64 layers because layer state tracking
used uint32_t, which only supports 32 bits (layers 0-31). Layers 32-63
could not be properly activated or tracked.

Changes:
- app/include/zmk/keymap.h: Changed zmk_keymap_layers_state_t from uint32_t to uint64_t
- app/src/conditional_layer.c: Changed then_layers and then_layer_state from uint32_t to uint64_t
- app/src/combo.c: Changed layer_mask in combo_cfg struct from uint32_t to uint64_t

This allows the layer state bitmask to handle all 64 layers (bits 0-63).

+++ app/src/combo.c
@@ -50,7 +50,7 @@ struct combo_cfg {
     int16_t key_position_len;
     int16_t require_prior_idle_ms;
     int32_t timeout_ms;
-    uint32_t layer_mask;
+    uint64_t layer_mask;
     struct zmk_behavior_binding behavior;
     // if slow release is set, the combo releases when the last key is released.
     // otherwise, the combo releases when the first key is released.
